chore: Rename nodeService to node-service
