+++ packages/neuron-wallet/src/services/addresses.ts
@@ -1,6 +1,6 @@
 import TransactionsService from './transactions'
 import WalletService from './wallets'
-import nodeService from '../startup/nodeService'
+import nodeService from '../startup/node-service'
 import HD from '../keys/hd'
 import { KeysData } from '../keys/keystore'
