+++ packages/neuron-wallet/src/startup/sync-block-task/create.ts
@@ -3,7 +3,7 @@ import { Subject } from 'rxjs'
 import path from 'path'
 import { networkSwitchSubject, NetworkWithID } from '../../services/networks'
 import app from '../../app'
-import nodeService from '../nodeService'
+import nodeService from '../node-service'
 import env from '../../env'
 import initConnection from '../../typeorm'
 import { AddressesUsedSubject } from '../../subjects/addresses-used-subject'
